requirements.txt
boto3
pycryptodomex
tabulate
pathvalidate
